chore: isolate auth client helpers
