+++ __tests__/lib/api/google-client.test.ts
@@ -1,31 +1,36 @@
-jest.mock('@/lib/api/base/api-client', () => ({
-  createApiClient: jest.fn((cfg) => { global.cfg = cfg; return {}; })
-}))
+jest.mock("@/lib/api/base/api-client", () => ({
+  createApiClient: jest.fn((cfg) => {
+    global.cfg = cfg;
+    return {};
+  }),
+}));
 
-jest.mock('@/lib/api/auth-interceptor', () => ({ wrapAuthError: jest.fn() }))
+jest.mock("@/lib/api/auth-errors", () => ({ wrapAuthError: jest.fn() }));
 
-jest.mock('@/lib/api/api-enablement-error', () => ({
+jest.mock("@/lib/api/api-enablement-error", () => ({
   isAPIEnablementError: jest.fn(() => true),
-  createEnablementError: jest.fn((e) => new Error('enable ' + e.message)),
-}))
+  createEnablementError: jest.fn((e) => new Error("enable " + e.message)),
+}));
 
-import '@/lib/api/google/client'
-import { APIError } from '@/lib/api/utils'
-import { wrapAuthError } from '@/lib/api/auth-interceptor'
-import { createEnablementError } from '@/lib/api/api-enablement-error'
+import "@/lib/api/google/client";
+import { APIError } from "@/lib/api/utils";
+import { wrapAuthError } from "@/lib/api/auth-errors";
+import { createEnablementError } from "@/lib/api/api-enablement-error";
 
-const passed = (global as any).cfg
+const passed = (global as any).cfg;
 
-describe('googleApiClient handleProviderError', () => {
-  it('wraps auth and enable errors', () => {
-    const err = new APIError('bad', 401)
-    expect(() => passed.handleProviderError(err)).toThrowError()
-    expect(wrapAuthError).toHaveBeenCalled()
-  })
+describe("googleApiClient handleProviderError", () => {
+  it("wraps auth and enable errors", () => {
+    const err = new APIError("bad", 401);
+    expect(() => passed.handleProviderError(err)).toThrowError();
+    expect(wrapAuthError).toHaveBeenCalled();
+  });
 
-  it('handles enablement error', () => {
-    const err = new APIError('not enabled', 403)
-    try { passed.handleProviderError(err) } catch {}
-    expect(createEnablementError).toHaveBeenCalledWith(err)
-  })
-})
+  it("handles enablement error", () => {
+    const err = new APIError("not enabled", 403);
+    try {
+      passed.handleProviderError(err);
+    } catch {}
+    expect(createEnablementError).toHaveBeenCalledWith(err);
+  });
+});
