+++ lib/api/google/client.ts
@@ -1,6 +1,9 @@
 import { createApiClient } from "../base/api-client";
-import { wrapAuthError } from "../auth-interceptor";
-import { createEnablementError, isAPIEnablementError } from "../api-enablement-error";
+import { wrapAuthError } from "../auth-errors";
+import {
+  createEnablementError,
+  isAPIEnablementError,
+} from "../api-enablement-error";
 import { APIError } from "../utils";
 import { getGoogleToken } from "../../steps/utils/auth";
 
@@ -8,7 +11,11 @@ import { getGoogleToken } from "../../steps/utils/auth";
  * Normalize Google API errors into authentication or enablement errors.
  */
 function handleGoogleError(error: unknown): never {
-  if (error instanceof APIError && (error.status === 401 || error.message?.includes("invalid authentication credentials"))) {
+  if (
+    error instanceof APIError &&
+    (error.status === 401 ||
+      error.message?.includes("invalid authentication credentials"))
+  ) {
     throw wrapAuthError(error, "google");
   }
   if (error instanceof APIError && isAPIEnablementError(error)) {
@@ -21,7 +28,7 @@ function handleGoogleError(error: unknown): never {
  * API client instance preconfigured for Google Workspace APIs.
  */
 export const googleApiClient = createApiClient({
-  provider: 'google',
+  provider: "google",
   getToken: getGoogleToken,
   handleProviderError: handleGoogleError,
 });
